WIP refactor the transforms
The overall idea is to express scan and detector coordinates
as such transforms, and allow to derive forward and backward transform
functions between physical coordinates and pixel coordinates
very simply and concisely for later use in the model.

In a later stage one can also use similar approaches for 3D coordinates
to express kinematics of complex actuators, such as specimen stages etc.
That was an issue with interpreting NeXus files from beamlines with such
actuators. At the time we were missing a
handy library to interpret the actuator positions in a meaningful way.
It could become very simple with a composition of transforms that
are validated against a model of the beamline, for exmple in TEMGYM.

Work to make them explicit in x and y coordinate for the user, similar to
the rays and components.

Base them on a simple definition of origin and base vectors since that might
be intuitive for users and also allows for sheared coordinate systems.

Not sure if this is a good approach, though... Just playing around and
seeing how it feels!

The missing transform functions can be added bit by bit from the original file.